djangofiles/BlogProject/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Django==2.2
django-clear-cache==0.3
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-role-permissions==2.2.1
entrypoints==0.3
flake8==3.7.7
flake8-django==0.0.4
importlib-metadata==0.18
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
Pillow==6.0.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.0
pytest==4.6.3
python-slugify==3.0.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
wcwidth==0.1.7
zipp==0.5.1
